Failure to learn on Asterix
See Apprentissage.ipynb. The score of the exploration policy is in blue, it gets stuck at 500 (less than the Const baseline in the JAIR paper). Maybe the NN of NFQ diverged or something.

The exploitation policy, in green, is no better.

I'm stopping here trying to work with asterix, I'll try Boxing instead because in this game the RAM features show the best performance w.r.t. the baseline and even other methods of RL.